Caching in Router
Idea:
1. Router has plenty of RAM, which can be used to cache hot keys;
2. With D2 sticky routing, Router could have an aggregated view of hot
keys, so that Router-layer caching will be much more efficient than Client caching;

Changes:
1. Introduced several new router configs:
router.cache.enabled : whether the cache is enabled on Router or not
router.cache.size.in.bytes : cache capacity in one Router
router.cache.concurrency : the concurrency count supported by cache, it
means how many requests at most can be executed in cache in parallel
router.cache.lookup.request.throttle.weight : the throttle weight for
cache hit request
2. Added store-level flag to enable/disable caching;
3. Implemented on-heap LRU cache with concurrency support in Router;
4. Moved the throttling logic for single-get from VeniceDelegateMode to
VeniceDispatcher because of the caching logic;

New per-store and aggregated metrics on router for this feature:
cache_lookup_request
cache_hit_request
cache_hit_ratio
cache_lookup_latency
cache_put_request
cache_put_latency

TODO:
1. Explore off-heap LRU cache to mitigate potential GC issue introduced
by large on-heap cache;
2. Add more unit tests;

RB=1168238
G=voldemort-reviewers
R=fvillega,yayan
A=fvillega

+++ venice-common/src/main/java/com/linkedin/venice/throttle/EventThrottler.java
@@ -131,7 +131,7 @@ public class EventThrottler {
    * @param eventsSeen Number of events seen since last invocation. Basis for
    *        determining whether its necessary to sleep.
    */
-  public synchronized void maybeThrottle(int eventsSeen) {
+  public synchronized void maybeThrottle(double eventsSeen) {
     if (maxRatePerSecond > 0) {
       long now = time.milliseconds();
       try {
